First Commit
Duck class is an abstract class and there are three sub-classes:
MallardDuck, RedheadDuck and RubberDuck.  Each class inherits Display(),
Quack(), Swim() and Fly().

Display is an abstract method, so each sub-class is forced to provide an
implementation.

Quack and Fly are virtual methods, so they have default implementations,
but can be overriden at the sub-class level.

Swim is an implemented method and cannot be overriden.